AI Engineer-  Assignment/requirements.txt
langchain
langgraph
langsmith
langchain-community
langchain-huggingface
qdrant-client
sentence-transformers
streamlit
pypdf
python-dotenv
requests
pytest
pytest-asyncio
typing-extensions
tenacity

